c++ Pathfinding Success
- Converted pathfinding to C++ and fixed the crash from the previous commit where it would crash on second Play.
- Also fixed other bugs and a possible memory leak from similar code.
- Made units not move at all until the pathing was calculated. Since the calculation is WAY faster now in C++ than it was in C#, this is hardly noticeable, if at all

Placement new destruction - 
Now: ::operator delete(memory)
Old: delete[] memory

+++ FlowField/IntegrationField.h
@@ -23,7 +23,7 @@ namespace Pathing
     protected:
 
     public:
-      unsigned short Distance;
+      unsigned int Distance;
       bool Used;// = false;
 
       IFieldNode(int x, int y, IntegrationField* i);
